refactor: merge v2 editor into EditView.vue
replaces old EditView with two-panel layout,
renames CollapsibleSectionV2 to CollapsibleSection,
removes v2 route and button from home
