+++ src/components/CollapsibleSection.vue
